Remove BookResult
