+++ src/OszkConnector/Client.cs
@@ -13,7 +13,7 @@ namespace OszkConnector
     {
         private const string MEK_ENDPOINT_URL = "http://mek.oszk.hu";
 
-        public async Task<IQueryable<BookResult>> FindAudioBook(string query = "")
+        public async Task<IQueryable<Book>> FindAudioBook(string query = "")
         {
             var uri = new Uri($"{MEK_ENDPOINT_URL}/katalog/kataluj.php3");
             var postData = $"szerint=cimsz&s1=dc_format+format_name&m1=MP3&muv1=and&s2=dc_creator_o+FamilyGivenName&m2={query}&muv2=or&s3=dc_title+subtitle&m3=&muv3=or&s4=dc_subject+keyword&m4=&muv4=or&s5=dc_subject+keyword&m5=&subid=&x=31&y=8&sind1=0&sind2=7&sind3=16&sind4=13&sind5=18&muv1index=0&muv2index=0&muv3index=0&muv4index=0&subid_check=&ekezet_check=&ekezet=ektelen";
@@ -25,7 +25,7 @@ namespace OszkConnector
             return books;
         }
 
-        public async Task<IQueryable<BookResult>> FindBook(string query = "")
+        public async Task<IQueryable<Book>> FindBook(string query = "")
         {
             return await MekCatalog.SearchAsync(query, "MP3");
         }
