+++ src/OszkConnector/Repository/IBookRepository.cs
@@ -8,9 +8,9 @@ namespace OszkConnector.Repository
 {
     public interface IBookRepository
     {
-        IQueryable<BookResult> GetAll();
+        IQueryable<Book> GetAll();
 
-        IQueryable<BookResult> Find(string query);
+        IQueryable<Book> Find(string query);
 
         Book Get(string Id);
     }
